requirements.txt
filelock==3.16.1
fsspec==2024.10.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
pandas==2.2.3
pillow==11.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
setuptools==75.5.0
six==1.16.0
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
typing_extensions==4.12.2
tzdata==2024.2
